requirements.txt
supabase
google-api-python-client
google-generativeai
feedparser
python-dotenv
